Fixed WWW error handling
From the documentation: https://docs.unity3d.com/ScriptReference/WWW-error.html

If there was no error, error will return null or an empty string (this is because some platforms don't allow nulls for string values). We recommend that you use String.IsNullOrEmpty to check for the presence of an error so that both cases are covered.